requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
dj-database-url==0.5.0
dj-rest-auth==6.0.0
dj3-cloudinary-storage==0.0.6
Django==5.0.6
django-allauth==0.63.1
django-autoslug==1.9.9
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
